First pass at consolidating general validations for HCA only
