+++ src/js/edu-benefits/utils/validations.js
@@ -27,13 +27,13 @@ function validateIfDirtyDateObj(date, validator) {
   });
 }
 
-function validateIfDirtyProvider(field1, field2, validator) {
-  if (field1.dirty || field2.dirty) {
-    return validator(field1.value, field2.value);
-  }
+// function validateIfDirtyProvider(field1, field2, validator) {
+//   if (field1.dirty || field2.dirty) {
+//     return validator(field1.value, field2.value);
+//   }
 
-  return true;
-}
+//   return true;
+// }
 
 function isBlank(value) {
   return value === '';
@@ -427,7 +427,7 @@ export {
   validateIfDirty,
   validateIfDirtyDate,
   validateIfDirtyDateObj,
-  validateIfDirtyProvider,
+  // validateIfDirtyProvider,
   initializeNullValues,
   isBlank,
   isNotBlank,
