+++ src/js/hca/components/household-information/SpouseInformationSection.jsx
@@ -8,7 +8,8 @@ import FullName from '../../../common/components/questions/FullName';
 import Phone from '../../../common/components/questions/Phone';
 import SocialSecurityNumber from '../../../common/components/questions/SocialSecurityNumber';
 import { yesNo } from '../../../common/utils/options-for-select.js';
-import { isNotBlank, validateIfDirty, isValidMarriageDate, isValidLastName } from '../../utils/validations';
+import { validateIfDirty, isNotBlank } from '../../../common/utils/validations';
+import { isValidMarriageDate, isValidLastName } from '../../utils/validations';
 import { veteranUpdateField, updateSpouseAddress } from '../../actions';
 
 // TODO: Consider adding question for marital status here so if user
