+++ src/js/hca/components/insurance-information/MedicareMedicaidSection.jsx
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import DateInput from '../../../common/components/form-elements/DateInput';
 import ErrorableRadioButtons from '../../../common/components/form-elements/ErrorableRadioButtons';
 import { yesNo } from '../../../common/utils/options-for-select';
-import { validateIfDirty, isNotBlank } from '../../utils/validations';
+import { validateIfDirty, isNotBlank } from '../../../common/utils/validations';
 import { veteranUpdateField } from '../../actions';
 
 /**
