+++ src/js/hca/components/insurance-information/VaMedicalFacility.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import _ from 'lodash';
 
 import ErrorableSelect from '../../../common/components/form-elements/ErrorableSelect';
-import { validateIfDirty, isNotBlank } from '../../utils/validations';
+import { validateIfDirty, isNotBlank } from '../../../common/utils/validations';
 import { vaMedicalFacilities } from '../../../common/utils/options-for-select.js';
 
 /**
