+++ src/js/hca/components/va-benefits/VAInformationSection.jsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 
 import ErrorableRadioButtons from '../../../common/components/form-elements/ErrorableRadioButtons';
 import { yesNo } from '../../../common/utils/options-for-select';
-import { validateIfDirty, isNotBlank } from '../../utils/validations';
+import { validateIfDirty, isNotBlank } from '../../../common/utils/validations';
 import { veteranUpdateField } from '../../actions';
 
 /**
